fix: define vm.box for virtualbox provider only
Previous commit added the vm.box to PXE booting VMs to be able to start
a domain with the virtualbox provider. However, this breaks the support
for PXE booting with the libvirt provider that requires no box.

With this change, the vm.box is added to VMs only when the provider is
virtualbox.