feat: create our own SSH key pair for root
Create the SSH key pair in .vagrant/ssh/ and authorized_keys in the
current working directory. Install the key pair in /root/.ssh/ on the
first management node and install the authorized_keys file on each
management node for root.

This allows SSH pubkey authentication from the first management node to
other management nodes which are provisioned from vagrant boxes.

Enforce the mode and ownership of the directory /root/.ssh with rsync.