fix: copy authorized_keys with file provisioner
vagrant-libvirt does not sync the current working directory to /vagrant
when the fs type is NFS, which break the copy of authorized_keys.
See: lib/vagrant-libvirt/action/share_folders.rb@5e774462#L28

Create the authorized_keys in .vagrant/ssh/ and copy it to first
management node with synced_folder, along with id_rsa and id_rsa.pub.
Copy authorized_keys to other management nodes with the file
provisioner.

Recreate the SSH key pair when one if the file is missing.