feat: set custom prefix to VM name
By default, the vagrant-libvirt provider defines the guest name prefix
from the name of the current working dir. It creates some conflicts when
one run the same profile with different configuration files on a single
host.

This change allows to define the default prefix in the vagrant.yml
configuration file, thus one can run the same profile with different
configuration on a single host.

For vagrant-libvirt only.